+++ parsedmarc/constants.py
@@ -1,2 +1,2 @@
-__version__ = "8.18.9"
+__version__ = "8.19.0"
 USER_AGENT = f"parsedmarc/{__version__}"